+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 
 ### Removed
 - Remove category `services_count` (@mkasztelnik)
+- Remove unnecessary logs in js console(@goreck888)
 
 ### Fixed
 - Hide scrollArrow on scroll event (@goreck888)