+++ CHANGELOG.md
@@ -20,6 +20,11 @@ Please view this file on the master branch, on stable branches it's out of date.
 
 ### Security
 
+## [2.1.5] 2019-10-09
+
+### Fixed
+- Fix IID in projects (@michal-szostak, @martaswiatkowska)
+
 ## [2.1.4] 2019-09-20
 
 ### Changed